# backend/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "models.AdminUser": {
                "properties": {
                    "avatar_url": {
                        "type": "string"
                    },
                    "blocked_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "is_admin": {
                        "type": "boolean"
                    },
                    "login": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "models.CreateUserRequest": {
                "properties": {
                    "display_name": {
                        "type": "string"
                    },
                    "is_admin": {
                        "type": "boolean"
                    },
                    "login": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "display_name",
                    "login",
                    "password"
                ],
                "type": "object"
            },
...
            "models.UpdateUserRequest": {
                "properties": {
                    "blocked": {
                        "type": "boolean"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "is_admin": {
                        "type": "boolean"
                    },
                    "login": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Заголовок вида `Bearer \u003cjwt\u003e` (получить через `/auth/login`).",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "name": "msdnna",
            "url": "https://github.com/msdnna/budget"
        },
        "description": "REST API семейного бюджет-приложения (Go + MongoDB). JWT (Bearer) — защищённые ручки требуют заголовок `Authorization: Bearer \u003ctoken\u003e`, полученный через `POST /api/auth/login`.",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "title": "msdnna Budget API",
        "version": "1.0"
    },
...
    "paths": {
        "/admin/users": {
            "get": {
                "description": "Возвращает login/is_admin/blocked_at/created_at в дополнение к публичным полям. Не-удалённые пользователи (включая заблокированных).",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/models.AdminUser"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Список пользователей с админ-полями (admin)",
                "tags": [
                    "users"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/models.CreateUserRequest",
                                        "summary": "body",
                                        "description": "Тело"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Тело",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.AdminUser"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "login already taken"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Создать пользователя (admin)",
                "tags": [
                    "users"
                ]
            }
        },
        "/admin/users/{id}": {
            "delete": {
                "description": "Помечает `deleted_at`. UserInfo-снимки в существующих записях остаются валидными. Нельзя удалить самого себя или последнего админа.",
                "parameters": [
                    {
                        "description": "User ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "boolean"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Soft-удалить пользователя (admin)",
                "tags": [
                    "users"
                ]
            },
            "patch": {
                "description": "Partial PATCH: login / display_name / is_admin / blocked. Safeguards: нельзя снять админку или заблокировать самого себя, нельзя снять админку с последнего админа.",
                "parameters": [
                    {
                        "description": "User ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/models.UpdateUserRequest",
                                        "summary": "body",
                                        "description": "Тело"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Тело",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.AdminUser"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Conflict"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Обновить пользователя (admin)",
                "tags": [
                    "users"
                ]
            }
        },
        "/admin/users/{id}/avatar": {
            "delete": {
                "parameters": [
                    {
                        "description": "User ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.AdminUser"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Удалить аватар (admin)",
                "tags": [
                    "users"
                ]
            },
            "post": {
                "description": "PNG / JPEG / SVG, до 512KB. Сохраняется inline в коллекции users.",
                "parameters": [
                    {
                        "description": "User ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "title": "file",
                                "type": "file"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    },
                    "description": "Файл",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.AdminUser"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Загрузить аватар пользователя (admin)",
                "tags": [
                    "users"
                ]
            }
        },
        "/admin/users/{id}/password": {
            "post": {
                "parameters": [
                    {
                        "description": "User ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/models.SetPasswordRequest",
                                        "summary": "body",
                                        "description": "Тело"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Тело",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "boolean"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Сменить пароль другого пользователя (admin)",
                "tags": [
                    "users"
                ]
            }
        },
...
        "/auth/password": {
            "post": {
                "description": "Требует старый пароль. После смены текущая JWT-пара остаётся валидной до истечения; рекомендуется перелогиниться.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/models.ChangePasswordRequest",
                                        "summary": "body",
                                        "description": "Тело"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Тело",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "boolean"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Старый пароль не подходит"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Сменить собственный пароль (self)",
                "tags": [
                    "auth"
                ]
            }
        },
...
        "/users/{id}/avatar": {
            "get": {
                "parameters": [
                    {
                        "description": "User ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "image/png": {
                                "schema": {
                                    "type": "file"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Отдать аватар пользователя (raw bytes)",
                "tags": [
                    "users"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api"
        }
    ]
}